[Refactor] Rename import
