+++ job_shop_lib/visualization/graphs/__init__.py
@@ -14,7 +14,7 @@ from ._plot_disjunctive_graph import (
     plot_disjunctive_graph,
     duration_labeler,
 )
-from ._plot_heterogeneous_graph import (
+from ._plot_resource_task_graph import (
     plot_resource_task_graph,
     three_columns_layout,
     color_nodes_by_machine,
